Add additional_rights to JSON output, fixes #1272
Our JSON output for a project didn't include additional_rights
information because that is not stored in the projects table
(it's in another table).  However, it's clearly relevant to a project.
Therefore, add the additional_rights field to the JSON output.
We'll output it as a JSON array (that seems like the best way to
model it).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>