Improve retrieval of user repo list
Improve the retrieval of the list of repos managed by a user.
Recent changes looked fine at first, but turned out to have
subtle problems.  In particular, the way we called Octokit
didn't pass parameters correctly, with the result that
in some cases we didn't get the list we were expecting.

This involved fixing/changing some test and test data to match.

In the process of fixing this, we also found that we were
unintentionally asking GitHub *twice* for the same data, which
is not good when you're trying to avoid a timeout :-).
So fix that as well.

Some places were unintentionally changing the *global* auto_paginate
value when retrieving data.  Change this to set the auto_paginate value
appropriately for the specific request being made, instead of doing it
globally.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>