Speed getting GitHub project list
Improve getting the GitHub project list.
When looking for a specific repo, halt immediately if it's found
instead of loading the whole list.
When showing a project, make the requests to GitHub (if needed) at most once.

Note that github_user_projects_include?(url) and
its callee repo_url_in? are not currently covered by a test.
That will need to be added.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>