Make "Account" the top header (constant display)
Change the visible top header so that it always just shows
"Account".  For now its dropdown still varies depending on
whether or not the user is logged in.
In the longer term we can use JavaScript to hide entries
(depending on whether or not the user is logged in as detected by a cookie;
if JavaScript isn't enabled, we'll just show all the options).
By putting them all *under* the "Account" there will be no flashing
or delay while waiting for JavaScript to hide some entries.

The eventual goal is to send exactly the same file
to everyone for certain pages such as the home page and single-project show.
If we can do that, then we can place those pages in the CDN,
resulting in a tremendous speedup.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>