Allow CORS access to individual user data
Allow access to .json data to individual users
by client-side JavaScript by adding it to the
CORS access control list.

This also adds tests to ensure that we force no caching
of user data.  We disable caching to greatly reduce the
risk of an unintentional leak of user data via a cache.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>