Update gem chartkick (2.3.5->3.2.0)
Chartkick 2.3.5 has a vulnerability (CVE-2019-12732),
so update to repaired component.

I am pretty confident that the BadgeApp cannot be attacked through this
vulnerability, because the only data we use on pages with Chartkick are
calculated numbers and dates from trusted sources. But since mistakes
are always possible, it's best to update vulnerable compoents to
deny attackers from gaining even a toehold.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>