CVE-2015-9284 in omniauth and update transitive dependencies
Counter CVE-2015-9284 in omniauth.

Unfortunately, at the time of this writing
omniauth is STILL vulnerable (!). There is a shim
by a third party that *does* fix it, so while I'm always wary of
such things, this appears to be the best available solution.
I don't know the person who created this shim,
but I reviewed the code and it looks okay.  I could do this:

~~~~
gem 'omniauth-rails_csrf_protection', '0.1.2' # Counter CVE-2015-9284
~~~~

But to provide a stronger guarantee that what I reviewed is what will
be loaded, I'm specifying a specific hash reference.  That's no
guarantee, but it does make attacks harder to perform.

The third-party countermeasure requires that all omniauth uses use POST,
and not GET, but we *always* did that so it's not a change for us.

In addition, this updates commit the transitive (indirect) gem
dependencies.  I had some trouble when I didn't do the update, and it's
time anyway, so I just updated instead of fighting it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>