Remove sudo commands from setup-heroku.sh
The bionic buildpack image runs as root and
does not have sudo installed.  Because we
are only installing heroku and running remote
commands in the deploy stage, it is okay to
run as root in this instance. It is not worth
the time to create a custom image with an
unprivileged user for the deploy step only.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>