Remove unnecessary workarounds for turbolinks
This commit removes unnecessary workarounds for turbolinks.
In particular, it removes
`app/helpers/disable_turbolinks_helper.rb`, calls to it,
and the design.md documentation about it.

It turns out that to make `/project_stats` work with turbolinks,
all that is needed is to add the following to its HTML head:

`<meta name="turbolinks-visit-control" content="reload">`

The result is simpler and easier to maintain software.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>