Document avatar security/privacy and add a little privacy
Document that avatar services are considered "related sites"
for purposes of security/privacy, and explain what they're all about.

In addition, add the experimental
`referrerpolicy="no-referrer"` attribute to the avatar image
as discussed in the
[Mozilla img documentation](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img)
to provide additional privacy.
We aren't required to do this by our policies, but in general we
try to maximize privacy while still providing our services.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>