Add word breaks and break-word to project table
The "projects" table is wide, and that's a problem for
narrow screens (especially on mobile).
Part of the problem is that URLs don't easily break,
and some project names use characters like underscores
that are not automatically word breaks.

To resolve this, add word break hints on project names and
URLs, and also add word-wrap: break-word on project names
(word-wrap is already set on URLs).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>