Omit the license field from /projects
The /projects page is wide, especially for mobile devices.
Few people really need to see the license at this level.
So remove it so that it's not as wide (or that other kinds of
information can be seen more readily).

The information is still available; users can always get license data
from an individual project, or download the database.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>