Add a "system announcement" capability
Add a "system announcement" capability.
If environment variables SYSTEM_ANNOUNCEMENT
(and optionally SYSTEM_ANNOUNCEMENT_locale) are set, show them on
every application page as an announcement.

The immediate purpose is so that we can warn about a
soon-to-occur brief shutdown for a database upgrade.
But this is a general-purpose mechanism that can be used for other
purposes as well.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>