HTTPS with GitHub Pages
This patch removes the sentence suggesting that people need to use a CDN
such as CloudFlare to enable HTTPS for GitHub pages when using custom
domains. This was true a while ago but is not any longer right now.

If a custom domain is configured for GitHub pages, now GitHub
automatically generates a Let's Encrypt certificate for this page and
even the enforcement of HTTPS is merely a checkbox in the repository
settings.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>