Add rake task change_owner -- PROJECT NEW_OWNER
Add a new rake task named "change_owner" to simplify the process
of changing the owner of a project. Local use is:

~~~~
rake change_owner -- PROJECT OWNER
~~~~

Where PROJECT is the project badge entry number and OWNER is the user id
of the new owner.

You can run this remotely (if authorized) as:

~~~~
heroku run --app production-bestpractices rake change_owner -- PROJECT OWNER
~~~~

We can already do this by running a SQL command, but an error
(such as forgetting a WHERE clause) could lead to a big mistake.
This task also shows related information (including how to revert it),
which reduces the risk of a mistake and simplifies recovery from one.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>