Allow repo_url mod if REPO_URL_CHANGE_DELAY days pass
Allow a project to modify its repo_url, but only after
REPO_URL_CHANGE_DELAY days (currently 180 days) since the last
time its repo_url changed.  This should reduce admin effort
while still countering a subtle attack.

We are trying to counter subtle attacks where
a project tries to claim the good reputation or effort of another project
by constantly switching its repo_url to other projects and/or nonsense.
The underlying problem is that names/identities are hard; the repo_url
(when present) is the closest to an "identity" that we have for a project.
We have to allow it to change sometimes (because it sometimes does), but
it should be a rare "sticky" event.

There are various special cases, e.g., you can always set the repo_url
if it's nil, the setter is an admin, or if only the scheme is changed.
But otherwise normal users can't change the repo_urls in less than
REPO_URL_CHANGE_DELAY days.

Allowing users to change repo_urls, but only
with large delays, reduces administration effort while still
countering the attack.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>