Tweak error messages/docs to support repo_url delay
Modify `config/locales/en.yml` so that the error message
clearly explains to users why they can't make the change
(if it is forbidden).

Modify `doc/security.md` to explain and justify the
`repo_url` modification delay.  This is security-related,
so we want to make sure we explain it!

Finally, tweak some comments to clarify this change.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>