Tweak criteria text involving CVSS
We want projects to focus on *important* vulnerabilities,
not "vulnerabilities" that aren't important.
To help, we use CVSS, but there are now multiple versions of CVSS
as CVSS gets updated.

This commit tweaks the criteria text (mainly in details) per the
mailing list discussion.  It's a lot of text, because
I couldn't figure out how to make it simpler.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>