Tweak: Modify account activation order
Modify app/controllers/account_activations_controller.rb
so that we *first* save the time logins can occur, and *then*
activate it. As-is there's a very tiny window of time
where the account is activated without a start time, which
would allow login. It'd be hard to *hit* that window in practice,
but best to eliminate it entirely.

Double-checked by Jason Dossett.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>