Modify mailer handling of 'X-SMTPAPI' header (#1388)
* Modify mailer handling of 'X-SMTPAPI' header

Modify how the mailers handle header 'X-SMTPAPI'.

This commit causes the app to *globally* set the default 'X-SMTPAPI'
value so that we *always* disable SendGrid's clicktracking.
We had done that in many cases, but because it wasn't the
default we missed a few cases. There's no reason to enable it,
so we just disable it always.

This commit now sets 'Send-at' *within* X-SMTPAPI.
Earlier it was an SMTP header, which is how email headers
are normally handled, but SendGrid ignores the header in that location
and wants the Send-at value hidden as a JSON value inside the
X-SMTPAPI header. That's weird, but since that's what they require,
we need to change the code to do that.

While we're at it, we'll shorten the code and add some
explanatory comments.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Change mailers to allow override

Unfortunately it's not possible to override the global default
set for mailers. Therefore, we need to set the headers on each
mail action. Modify the code to do so, so that everything now
works as intended.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>