Modify new_activation_link_created to note delay
We've intentionally added a short delay after signing in before
the user receives the activation link.

This commit specifically mentions that delay to the user, so that the
user won't be surprised by that.  It also mentions why the delay
(it's an anti-spam measure). That should reduce the annoyance level of
real users, and it may help discourage spammers.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>