Refactor X-SMTPAPI and add tests for it
Sendgrid uses X-SMTPAPI in ways we depend on, in particular,
we want to forcibly disable clicktracking and ask for a
delay in sending activation emails.

This refactors the calculation of X-SMTPAPI to make it
easier to query/test, and adds tests for the value of
X-SMTPAPI in some sample emails.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>