Document how we handle omniauth and CVE-2015-9284
This documents in security.md how we handle omniauth and CVE-2015-9284
I want to remove the GitHub warning about omniauth and CVE-2015-9284,
but before I do, I think it's important to document the issue in
security.md so people won't think we're just ignoring it.
It's also a good example of the complications of the real world :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>