Document ActionController::TestCase is obsolete (#1403)
ActionController::TestCase is obsolete. The Rails developers want
people to switch to using ActionDispatch::IntegrationTest ; see:
https://github.com/rails/rails/issues/22496

However, this isn't immediately obvious in the code.
We include rails-controller-testing which keeps those tests running.

This commit clearly documents that as a "TODO" we want to
to switch from ActionController::TestCase to ActionDispatch::IntegrationTest ;
this will discourage people from adding tests using the
obsolete interface. It also encourages people to switch tests
to the newer system. Finally, it notes that once we remove all uses of
ActionController::TestCase we can remove the gem rails-controller-testing.

This merely commits comments, instead of "doing" anything.
But clearly documenting what needs to change, right where the code
needs to change, seems like a good first step to actually changing things.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>