Minor updates of many gems
Modify Gemfile (and the corresponding Gemfile.lock) to
do non-major updates of many gems.

This tweaks test/test_helper.rb to eliminate the noise during
testing declaring "Capybara starting Puma..."; the noise made
it harder to see overall test progress. The noise was added by
the update, so we needed to add a configuration statement to
silence it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>