Update gem vcr (4.0.0->5.0.0)
Update gem vcr.

I am intentionally *not* updating to version 5.1.0;
that version is released under the MIT-Hippocratic-1.0 license.
(Our license checking mechanisms flag this sort of thing.)

The Open Source Initiative (OSI) has definitively ruled that the
MIT-Hippocratic-1.0 license is *NOT* an open source software license.
Similarly, the FSF has determined that it is not a Free software license.
More discussion here:
https://itsfoss.com/hippocratic-license/

Our policy is to only use OSS for key portions.
I don't see a strong reason to make an exception here.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>