Update Rails (5.2.2.1->5.2.4.1)
This updates the Ruby on Rails web application framework
that we use as a basis for this application.
This application currently uses the "5.2" branch of Rais;
this commit updates to the current version of the Rails 5.2
branch, version 5.2.4.1, released 2019-12-18 (about 2 months ago).

I first reviewed all the CHANGES in the respective CHANGELOGs here:
https://weblog.rubyonrails.org/2019/12/18/Rails-5-2-4-1-has-been-released/
and after the update all tests passed.

This version of Rails fixes CVE-2019-16782 as described here:
https://www.openwall.com/lists/oss-security/2019/12/18/2
This application doesn't store session ids this way, so it is
immune to CVE-2019-16782. That said, it's best to be up-to-date so
we can rapidly update if another vulnerability is found that *might*
affect this applcation.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>