ProjectStatsControllerTest: Use IntegrationTest
ActionController::TestCase is deprecated.
Rails encourages switching to ActionDispatch::IntegrationTest,
which calls the router (which then calls the controller) instead
of directly calling the controller. IntegrationTest is preferred
because it is a somewhat more thorough test
(since it goes through the router, just like the real thing).
Historically ActionController::TestCase was faster, but that
is no longer the case.

This commit changes ProjectStatsControllerTest to use
ActionDispatch::IntegrationTest instead of ActionController::TestCase.
The goal is to improve our tests to be more thorough *and*
remove deprecated constructs.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>