Check password_resets inputs more cautiously
In a few cases the password_resets controller could be
forced to attempt a nil dereference, which would then raise
an exception. This was noticed in the process of
improving the tests for the password_resets controller.

In some languages (like C) that would be a
disaster. In Ruby on Rails, this simply leads to a 404
("not found") error.  This is one of many examples where
an environment that is "more safe by default" helps out.

Still, it'd be better to directly do input validation
on password resets to counter this sort of thing.
This commit does more careful error checking, and adds
corresponding tests.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>