Update gem puma (4.3.2->4.3.3)
This updates our webserver, puma.
In particular this addresses vulnerability CVE-2020-5247,
"HTTP Response Splitting vulnerability in puma".
For more information see:

* https://github.com/puma/puma/security/advisories/GHSA-84j7-475p-hp8v
* https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-5247

More detailed text about the vulnerability:
"If an application using Puma allows untrusted input in a response header,
an attacker can use newline characters (i.e. CR, LF or/r, /n) to end
the header and inject malicious content, such as additional headers
or an entirely new response body. This vulnerability is known as HTTP
Response Splitting."

The BadgeApp is *not* vulnerable to this vulnerability
(unless something else has gone terribly wrong).
We consider allowing untrusted input in a response header, without
seriously limiting whitelists, to be a vulnerability in itself.
After all, why should we allow an attacker to control our headers?!

That said, it's best to have multiple layers of defenses where
that is pragmatically reasonable. This way, even if we have a made
terrible mistake in our input validation, the webserver will be
defended against this kind of attack. So we're updating quickly,
even though we don't think we *need* to.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>