Update UsersControllerTest
Update UsersControllerTest to use ActionDispatch::IntegrationTest
instead of the obsolete ActionController::TestCase.

This changes all of its tests to use the router instead of
calling the controller directly, with the result that the tests
are more like the "real thing". I also added a number of
checks to ensure that the test didn't merely *run* but also
produced expected *results*.  In process it was revealed that
a case wasn't covered: if a user deletes himself, he should
automatically be logged off. This doesn't seems to be a *security*
problem (since the "non-user" would have no rights), but it
would leave the user in a weird state, so let's fix that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>