Update rails (5.2.4.1->5.2.4.2)
Update rails to fix a vulnerability in ActionView
(a component of Rails).

Report is GHSA-65cv-r6x7-79hv, moderate severity,
Vulnerable versions: < 5.2.4.2
Patched version: 5.2.4.2

I believe we are *not* vulnerable, as we do not directly use the
ActionView JavaScript literal escape helpers that are vulnerable.
(<%= j ... %> or escape_javascript). However, though unlikely,
it's possible that a library might use it, and in general we try to
NOT have known-vulnerable components in our system.

==

Quotation:

There is a possible XSS vulnerability in ActionView's JavaScript literal escape helpers. Views that use the j or escape_javascript methods may be susceptible to XSS attacks.

Impact:
There is a possible XSS vulnerability in the j and escape_javascript methods in ActionView. These methods are used for escaping JavaScript string literals. Impacted code will look something like this:

<script>let a = `<%= j unknown_input %>`</script>

or

<script>let a = `<%= escape_javascript unknown_input %>`</script>

"Thanks to Jesse Campos from Chef Secure"

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>