Update projects_controller_test.rb
Switch from ActionController::TestCase to
ActionDispatch::IntegrationTest.

A few tests are hard to convert, because the IntegrationTest
cannot manipulate the session, as documented in:
https://stackoverflow.com/questions/37796129/
unable-to-set-session-hash-in-rails-5-controller-test

So for the moment, those few tests are *not* converted and
are instead in the new file projects_controller_special_test.rb.
This means we still have a use of the deprecated
class ActionController::TestCase, but it is very small,
and that means that the rest of the tests are using
ActionDispatch::IntegrationTest (which has more realistic
tests, because it also tests the routing system).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>