Check GitHub permissions using repos API; Fixes #1418 (#1419)
For users with a lot of repositories, paging through
them can cause a timeout. This merge request uses a
more direct way to check repo permissions, the
GitHub repos API.

Additionally, To minimize calls to github also check
if a GitHub user owns the repo of a badge, if so
allow the user to edit the badge regardless of badge
project "owner."

Our changes produced more complicated test cases
that brought to light some n+1 queries, causing
a lot exceptions from the bullet gem. Those are
fixed now with a more generous usage of includes.

It is possible, though not tested, that this commit 
could could help with the problems in #1254.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>