Record when badges are *first* achieved
Record when each project's badge level is *first* achieved.

We currently record for each project the *last* time a
level was achieved or lost. However, it might be useful
to know when it was *first* achieved, so let's record that
*now* while the information is easier to get.
We'll estimate current projects' values by setting
first_achieved_passing_at = achieved_passing_at
WHERE achieved_passing_at IS NOT NULL.
We can add indexes and searches later if that's useful.

This also tweaks how times are recorded so that the
datatime is the same for all relevant achievement/loss
events, since there are now multiple relevant datetimes
that could be recorded for a given event.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>