Report achievement of higher badge levels (#1421)
* Report achievement of higher badge levels

In the monthly report, report projects that achieve
higher-level badges.

This was a classic "this will be an easy task" that turned
into more work than expected:
* We never recorded achieving
  higher-level badges, so we needed a data migration to
  add this information.
* I wanted the field names to be consistent, so we
  needed a method to convert level ids/numbers
  into field names.
* The previous step made it obvious that using
  Criteria.keys() as the way to find level IDs wasn't really
  sensible; that should be its own constant, and once that's done,
  modifying a few routines like update_prereqs() to use it seemed
  appropriate (the new version is clearer than the old one - at
  least Rubocop thinks so!).
  The whole thing should be clearer & thus less likely to have
  defects (it should be easier to add another level, too).
* The CSV statistics file generator missed some information,
  since we added new fields, and in general easily went wrong.
  Instead of manually adding fields, I rewrote the CSV template
  so that it automatically includes all stats.
  That completely future-proofs the whole thing; we may never
  need to modify it again.
* I had to temporarily disable Railroader again :-(
* A number of tests related to this needed to be modified.
  I added a number of extra tests to focus on the new
  parts I created.

This does have a change. It reports projects that achieve
a badge level in a month, even if they've lost it before and
the re-achieved it.  We want to encourage projects to get and
keep badges; if a project re-earns a badge they've lost, it
seems appropriate to congratulate them & note it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Record when badges are *first* achieved

Record when each project's badge level is *first* achieved.

We currently record for each project the *last* time a
level was achieved or lost. However, it might be useful
to know when it was *first* achieved, so let's record that
*now* while the information is easier to get.
We'll estimate current projects' values by setting
first_achieved_passing_at = achieved_passing_at
WHERE achieved_passing_at IS NOT NULL.
We can add indexes and searches later if that's useful.

This also tweaks how times are recorded so that the
datatime is the same for all relevant achievement/loss
events, since there are now multiple relevant datetimes
that could be recorded for a given event.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Restore railroader in CI process

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>