Update rails (5.2.4.2->5.2.4.3)
Update the Rails framework from 5.2.4.2->5.2.4.3.

This resolves the following potential vulnerabilities in the
underlying framework:
* [CVE-2020-8162] Circumvention of file size limits in ActiveStorage
* [CVE-2020-8164] Possible Strong Parameters Bypass in ActionPack
* [CVE-2020-8165] Potentially unintended unmarshalling of
  user-provided objects in MemCacheStore and RedisCacheStore
* [CVE-2020-8166] Ability to forge per-form CSRF tokens
  given a global CSRF token
* [CVE-2020-8167] CSRF Vulnerability in rails-ujs

All but one of these wouldn't work anyway. The CSRF token forging isn't
good, but we also use same-site cookies so you'd have to simultaneously
defeat both mechanisms to have a working attack.

The only that *might* have worked was [CVE-2020-8164] Possible
Strong Parameters Bypass in ActionPack. We don't know of any cases
where this would have *actually* led to a vulnerability, but it's
easier to simply upgrade and eliminate the problem instead of
wasting a lot of time to determine if we need to update.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>