Update dependencies
Update dependencies, in particular, update transitive dependencies
using `bundle update` and also update direct dependency
`pry-byebug` (3.8.0->3.9.0) because otherwise an undefined
constant for `pry` will cause the application to halt.

The key reason for this update is to update
websocket-extensions from 0.1.4 to 0.1.5 to address
vulnerability GHSA-g6wq-qcwm-j5g2
(moderate severity, vulnerable versions: < 0.1.5, patched version: 0.1.5).

This was just reported within the past hour.
I don't think we're vulnerable to this.
Also, this is merely a denial-of-service attack, so
it cannot corrupt our data if we are attacked this way.
That said, it's safest to update and avoid the issue.

Here are the vulnerability details:

> The ReDoS flaw allows an attacker to exhaust the server's capacity to process
incoming requests by sending a WebSocket handshake request containing a header
of the following form:

> Sec-WebSocket-Extensions: a; b="\c\c\c\c\c\c\c\c\c\c ...

> That is, a header containing an unclosed string parameter value whose content is
a repeating two-byte sequence of a backslash and some other character. The
parser takes exponential time to reject this header as invalid, and this will
block the processing of any other work on the same thread. Thus if you are
running a single-threaded server, such a request can render your service
completely unavailable.

> Users should upgrade to version 0.1.5.

> https://blog.jcoglan.com/2020/06/02/redos-vulnerability-in-websocket-extensions/

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>