Add Argon2id from NIST Password Hashing Competion
5 years ago, Argon2 won NIST's Password Hashing Competion.
It is included in OWASP's recommendations:
 https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html
I think it has seen sufficient experience and vetting to add to this
list.
I also put PBKDF2, which is not memory-hard, at the end of the list.
It seems to still be around until some other algorithm is approved
for FIPS compliance last I heard.

It seems most helpful to name Argon2id specifically since it is called
out in the (still-draft) IETF ID:
 https://tools.ietf.org/html/draft-irtf-cfrg-argon2-10.

Signed-off-by: Neal McBurnett <neal@mcburnett.org>