Move edits to en.yml; authorization token note
As noted by david-a-wheeler, criteria.md is auto-generated from en.yml.
(Should it even be version-controlled?)

Also changed another place where bcrypt is mentioned.

And there is a comment about passwords for logging in to other sites.
Do we really want to condone that? I've changed it to "authentication tokens",
and guess that oauth or the like might be worth mentioning, but that's a whole
new topic, I guess.

Signed-off-by: Neal McBurnett <neal@mcburnett.org>