Purge CDN last
This commit changes CDN purging of badge data to be last.

We've received two reports where an edit of project data worked,
but the badge image display has the old value.
Jason Dossett & I believe the problem is some sort of race condition
with Fastly, but we've not been able to reproduce it.

This moves CDN purges to be the last task done, which should
reduce the likelihood of triggering the race condition.
Before, in the save-after-edit case, there was nontrivial functionality
after a badge purge, making race conditions more likley.

This commit doesn't actually *fix* the race condition, but it should
significantly reduce the likelihood of it occurring, and that's a good thing.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>