Send CDN purge early, as well as late
Send CDN purge early to counter potential delays within the CDN.

When the project is about to be edited, send the CDN a purge request
early as well as late. This increases the likelihood that a purge
will complete before a request is sent if it takes significant time
for the CDN to distribute that information through its internal network.

We continue to send the purge request late, so that if the CDN manages
to get in a request after the early purge but before the change occurs,
the CDN will throw it away.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>