Add presentation doc/cii-bp-badge-intro.pptx
This commit adds a presentation about the badge in .pptx format
(it's in .pptx format because that's what the original was in).
This makes it easier for people to quickly get this presentation.
It has a relatively small number of slides that provide an overview,
combined with many backup slides to provide more information on various cases.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>