Remove year in per-file copyright statement
Remove the requirement for indicating a year in the copyright date.
For a much longer discussion, see:
https://github.com/coreinfrastructure/best-practices-badge/pull/1439

This was originally proposed by John Mertic ( @jmertic ) on discussion
with @swinslow and many others. This is basically a minor
relaxation of a gold badge requirement. It was proposed nearly a month
ago, and while there's been sporatic discussion to word it "just right"
there have been no objections to the change.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>