Fix test locales
The tests are *supposed* to use English locale unless otherwise
noted, but test 'ProjectStat.percent_field_description() works correctly'
does not do so consistently. This is suddenly a serious problem,
because it's preventing easy updating of a component with a
potentially serious vulnerability.

We'll resolve it by forcing the :en locale in the flapping test, and also
by reordering test 'Ensure tiered_percent_as_string works'
(the order *shouldn't* matter, but it might not be reliably cleaning up
after itself, so reordering it to reduce the risk seems reasonable).

There's nothing terribly wrong about these changes, and we can look
for more elegant solutions later. Right now I want to upgrade the
vulnerable component *with* confidence in the tests.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>