Improve German translation (merkrafter) fixes #1423
Incorporate a large number of improvements to the German translation
created by @merkrafter. Thank you very much!!

This fixes typos, changes the German text to be
consistent about "Sie" and "du", and corrects compound spelling.

This fixes #1423.

The changes were originally made in the YAML file, so we had to work out
how to transmit YAML changes to translation.io (which I've done).
That code will be added in a *separate* commit.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>