Change whitelist->allowlist,blacklist->denylist, fixes #1447 (#1449)
* Change whitelist->allowlist,blacklist->denylist, fixes #1447

Change occurrences of "whitelist" to "allowlist" and
"blacklist" to "denylist". NOTE: "a whitelist" becomes
"an allowlist" (note the change from "a" to "an").

This *is* a change in text of the criteria, which is normally a big
deal, but this change does *not* change the *meaning* of the criteria
in any way.

Translators will have to determine if they should change the
corresponding terms in their natural languages (or not).
I simply don't know enough about other languages' terms and cultures.

Please note that we *cannot* change the term "whitelist" in
`dependency_decisions.yml`. That is a machine-processed file and
the term is part of the API of the externally-implemented
tool `license_finder`. To be fair, this file is also extremely obscure.

This commit does *not* change the name of the master branch.  The
expectation is that we'll eventually rename the master branch, but I'm
waiting for more public guidance / agreement on a good replacement term.
My current preference is "main", and there are others who agree with me.
However, changing branch names can interfere with a *lot* of processes.
Thus, I'm waiting so that the branch rename will only occur *once*,
which should reduce disruption.

The term "allowlist" seems to be far more common, and
we prefer common terms. Thus, we'll use "allowlist" not "acceptlist"

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>