Update gem license_finder (5.6.2->6.0.0), rm "whitelist"
Update the gem license_finder from 5.6.2->6.0.0.

This removes the term "whitelist" from input file
`doc/dependency_decisions.yml`, as license_finder version 6.0.0
switches terms.

We had wanted to upgrade earlir, but we weren't able to do that because
of conflicting transitive dependency requirements on the gem thor
via bundler-audit. That's been resolved, so now we can move forward.

Note: version 6.0.0 is *not* the latest version of license_finder;
other transitive dependencies (e.g., on thor) prevent that.
We may be able to move to later versions of license_finder, but
I wanted to first transition to a new major version, as that is
less jarring. Incremental upgrades reduce the risk of surprises.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>