Update gem mdl, license permissions
The primary purpose is to update gem mdl (0.9.0->0.10.0).

This change led to the need to include the gems
chef-utils and mixlib-shellout, which are licensed with Apache 2.0.
This led to more license review requirements.

This commit also eliminates a source of unnecessary license problem reports.
Apache 2.0 wasn't on the "always permitted list" because there's
a potential issue with GPL-2.0-only and Apache 2.0.
But GPL-2.0-only is quite rare; important projects like the Linux kernel
and git use this license, but most GPL licenses use GPL-2.0-or-later.
So while we need to continue to ensure that everything is legal,
this commit takes the simpler path: allow "Apache 2.0",
"GPL-2.0-or-later", and "GPL-3.0-only", as they're all known to
be compatible. The license_finder will *still* correctly warn about
GPLv2-only software (since that has to be checked for how it's used),
So we'll continue to check for potential issues, but this tweak will
eliminate a common cause of unnecessary license reports.
I view this tweak as a minor internal tweak of our processes
(it just implements our current policies better), so I'm merging it
directly into the master (main) branch so that these library updates
can be smoothly added.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>